chore: interface

+++ src/Bases/Auctioneer/BaseAuctioneer.sol
@@ -169,7 +169,7 @@ abstract contract BaseAuctioneer is BaseHealthCheck, ReentrancyGuard {
             auction.fromInfo.tokenAddress,
             want(),
             auction.kicked,
-            auction.kicked + auctionLength > block.timestamp
+            auction.kicked + uint256(auctionLength) > block.timestamp
                 ? auction.currentAvailable
                 : 0
         );
@@ -188,7 +188,7 @@ abstract contract BaseAuctioneer is BaseHealthCheck, ReentrancyGuard {
         returns (uint256)
     {
         // If not enough time has passed then `kickable` is 0.
-        if (auctions[_auctionId].kicked + auctionCooldown > block.timestamp) {
+        if (auctions[_auctionId].kicked + uint256(auctionCooldown) > block.timestamp) {
             return 0;
         }
 
@@ -433,7 +433,7 @@ abstract contract BaseAuctioneer is BaseHealthCheck, ReentrancyGuard {
         address _fromToken = auctions[_auctionId].fromInfo.tokenAddress;
         require(_fromToken != address(0), "not enabled");
         require(
-            block.timestamp > auctions[_auctionId].kicked + auctionCooldown,
+            block.timestamp > auctions[_auctionId].kicked + uint256(auctionCooldown),
             "too soon"
         );
 
@@ -516,7 +516,7 @@ abstract contract BaseAuctioneer is BaseHealthCheck, ReentrancyGuard {
         AuctionInfo memory auction = auctions[_auctionId];
         // Make sure the auction is active.
         require(
-            auction.kicked + auctionLength >= block.timestamp,
+            auction.kicked + uint256(auctionLength) >= block.timestamp,
             "not kicked"
         );
